Remove eos-administration submodule
eos-administration has tighter access controls, so it's very likely that
the submodule won't be able to be cloned. It's not actually needed since
all of the real setup and deployment code has been moved there. The
checkout here is only needed for developer testing. In that case, simply
clone the repo on demand when needed.

https://phabricator.endlessm.com/T22793