ping: Send the records to Redis queues
With the new metrics server architecture, another component will be
responsible for pulling from these Redis queues and storing the metrics
in a database.

The whole "configuration" thing is completely removed, as it will be
moved to that new system.

However, we do not automatically get the createdAt/updatedAt times from
the database abstraction any more, so we need to add them back
explicitly.