test: Fix handling of the promise
The skip_done was only necessary to avoid the done() function being run
twice.

But then it was not supposed to be run in the hook handler in the first
place, so removing it from there allows us to get rid of the whole
skip_done thing and have this single test be closer to all the other
ones.