deps: Stop requiring underscore
There were a few problems with this dependency.

First, it was only ever used in the tests, but it was declared as an
install dep instead of a development deps.

Second, it was used only once, for something which can trivially be done
without it.

All in all, let's remove it and make things leaner.