ping: Don't (attempt to) log full geoip lookup
What we are trying to do here is to reduce the IP address, using only an
offline database lookup, to a country code, and not store any more
specific location information.

Unfortunately, we previously would attempt to log the full result of the
geoip lookup! Fortunately, concatenating a string and an object doesn't
serialize the contents of the object, so all that would be logged is:

    info: Geo: [object Object]                                                                                                                                                                                                                                                                                                        |
Remove this useless log message.

https://phabricator.endlessm.com/T34717