Use ES6 import syntax, not CommonJS require()
This is necessary to update to Chai 5.x which dropped support for being
imported into CommonJS code, since ES6 modules are the more modern,
standardized module system.

This is all mostly boilerplate changes.

The one surprising change is that activation.unit.js and ping.unit.js used to
require() the module under test inside the describe() callback, rather than at
the top level. I can't see that moving it to the top level makes any
difference.

I think the two modules are wrong to have a top-level `redis` variable that is
set by a call to a function because if you call that function twice with two
different values you are changing the variable used by the routes returned from
the first call. But when they are imported is not relevant; it's the call to
the “class” (which in both cases is actually a factory function, and an arrow
function at that, not a class in the sense of the `class` keyword or of a
prototype-based function called with `new Foo()`) that sets the top-level
variable.