api: Add /health endpoint
Previously, the HEALTHCHECK configured in the Dockerfile would just make an
HTTP request to / and check that the result code is in the range 200–499. In
practice the result code would be 404 because there is no route at the root.

However, this does not actually check whether the service is healthy. The only
function of the service is to turn HTTP requests into entries in Redis queues.

Add a new /health route that increments a Redis key. This is done rather than
using PING because PING succeeds when we are connected to a read-only replica,
but for the service to be healthy it must be able to write to Redis.

Adapt the healthcheck script to hit that endpoint.

https://phabricator.endlessm.com/T32315