Improve error handler middleware
Passing a 4-argument function to `app.use` indicates that it is an error
handler, called when an exception escapes a request handler.

Previously, the function checked whether its first parameter, `err`, is truthy.
According to the documentation this is not necessary.

Previously, the function would always send a response. According to the
documentation, you must first check whether the response headers have already
been sent, and if so chain to the next handler (which will in practice be the
default handler that abruptly closes the connection in that case).

Previously, wheen running under NODE_ENV=test, the function would stringify the
error and then join it to the error's stack trace. This is unnecessary because
err.stack starts with err.toString(). Simplify this.

https://expressjs.com/en/guide/error-handling.html