redis: Reconnect on READONLY errors
When an ElastiCache Redis cluster is upgraded, the old nodes stick around as
long as someone is connected to them, but as readonly replicas. Attempting any
write operation fails with:

> READONLY You can't write against a read only replica

(As described in the ioredis documentation, this can also happen when a replica
is promoted to primary due to some error if the Auto-failure ElastiCache option
is disabled; but the ElastiCache clusters in Endless's instance of this service
have this option set to true, which leads to different behaviour.)

Handle the READONLY case as suggested in the ioredis documentation: implement a
callback for Redis protocol-level errors that matches against the prefix
READONLY and triggers a reconnect in this case.

(I would like to add an integration test for this but it's quite non-trivial to
do so...)

https://github.com/redis/ioredis/blob/main/README.md#reconnect-on-error
https://phabricator.endlessm.com/T35146