redis: Don't resend command after reconnecting due to READONLY
I made a local change to reconnectOnError() to return `2` to test the
behaviour. I then described in the comment why I thought that it would be
better not to do that, but forgot to change the `2` back to `1` or `true`.

Fixes commit a755997df1e0f654b76db29954816de3be150802.

https://phabricator.endlessm.com/T35146